+++ src/include/model.h
@@ -7,11 +7,11 @@ typedef struct
 {   
     layer*** inLayers;      // References to the input layers of the model - entry point for model operations, array of pointers to layer allocation pointers
     layer*** layerList;     // List of all layers in the model - for operations to encourage spatial locality, array of pointers to layer allocation pointers
-    layer **outLayer;       // References the output layers of the model - entry point for model operations - pointer to layer allocation pointer
+    layer** outLayer;       // References the output layers of the model - entry point for model operations - pointer to layer allocation pointer
     float *targets;         // target values for the current training iteration
     float learningRate;    // Learning rate for the NN
-    int numLayers;          // Number of total layers in the NN
-    int numInLayers;        // Number of input layers in the NN
+    size_t numLayers;          // Number of total layers in the NN
+    size_t numInLayers;        // Number of input layers in the NN
     char loss_fn;
 } model;
